// web/src/api/schema.d.ts
export interface components {
    schemas: {
// ...
        SearchCatalogResponse: {
            /** Books */
            books: components["schemas"]["CatalogBookResponse"][];
        };
// ...
        SearchResponse: {
            /** Book Counts */
            book_counts?: components["schemas"]["BookResultCountResponse"][];
            /** Cached */
            cached: boolean;
            /** Duration Ms */
            duration_ms: number;
            pagination?: components["schemas"]["SearchPaginationResponse"] | null;
            /** Query */
            query: string;
            /** Result Count */
            result_count: number;
            /** Results */
            results: components["schemas"]["SearchResultResponse"][];
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}